requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
boto3==1.17.30
botocore==1.20.30
cachetools==4.1.1
cchardet==2.1.7
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastprogress==1.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Dance==3.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.1
hyperas==0.4.1
hyperopt==0.2.5
idna==2.10
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
keras-bert==0.86.0
keras-embed-sim==0.8.0
keras-layer-normalization==0.14.0
keras-multi-head==0.27.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.46.0
keras-transformer==0.38.0
kiwisolver==1.3.1
ktrain==0.26.1
langdetect==1.0.8
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.5
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.95
seqeval==0.0.19
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.2
syntok==1.3.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
transformers==4.3.3
uritemplate==3.0.1
urllib3==1.26.2
URLObject==2.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.2.1
zipp==3.4.0
